Implementing audio oscillator

+++ core/src/cpp/include/host/audio/config.hpp
@@ -23,8 +23,8 @@ namespace MC64K::StandardTestHost::Audio::IConfig {
 
     int32   const PROCESS_RATE  = 48000;
     size_t  const PACKET_SIZE   = 256;
-    float32 const SAMPLE_PERIOD = 1.0f / (float32)PROCESS_RATE;
-    float32 const PACKET_PERIOD = (float32)PACKET_SIZE / (float32)PROCESS_RATE;
+    float64 const SAMPLE_PERIOD = 1.0f / (float64)PROCESS_RATE;
+    float64 const PACKET_PERIOD = (float64)PACKET_SIZE / (float64)PROCESS_RATE;
 };
 
 #endif
